trading_service/requirements.txt
fastapi
uvicorn
requests
pandas
fyers-apiv3
python-dotenv
selenium
webdriver-manager
fastapi 
uvicorn